chore: Resolve code review issues

+++ packages/vitest/src/api/setup.ts
@@ -40,6 +40,11 @@ export function setup(vitestOrWorkspace: Vitest | WorkspaceProject, server?: Vit
     })
   })
 
+  function checkFileAccess(path: string) {
+    if (!isFileServingAllowed(path, getServer()))
+      throw new Error(`Access denied to "${path}". See Vite config documentation for server.fs.`)
+  }
+
   function setupClient(ws: WebSocket) {
     const rpc = createBirpc<WebSocketEvents, WebSocketHandlers>(
       {
@@ -76,8 +81,7 @@ export function setup(vitestOrWorkspace: Vitest | WorkspaceProject, server?: Vit
           return ctx.snapshot.resolveRawPath(testPath, rawPath)
         },
         async readSnapshotFile(snapshotPath) {
-          if (!isFileServingAllowed(snapshotPath, getServer()))
-            throw new Error(`Reading of snapshot "${snapshotPath}" disallowed. See Vite config documentation for server.fs.`)
+          checkFileAccess(snapshotPath)
           if (!existsSync(snapshotPath))
             return null
           return fs.readFile(snapshotPath, 'utf-8')
@@ -93,10 +97,12 @@ export function setup(vitestOrWorkspace: Vitest | WorkspaceProject, server?: Vit
           return fs.writeFile(id, content, 'utf-8')
         },
         async saveSnapshotFile(id, content) {
+          checkFileAccess(id)
           await fs.mkdir(dirname(id), { recursive: true })
           return fs.writeFile(id, content, 'utf-8')
         },
         async removeSnapshotFile(id) {
+          checkFileAccess(id)
           if (!existsSync(id))
             throw new Error(`Snapshot file "${id}" does not exist.`)
           return fs.unlink(id)
